move types to types.ts
